+++ src/App/Fossa/Config/Container/Test.hs
@@ -3,7 +3,7 @@
 module App.Fossa.Config.Container.Test (
   ContainerTestConfig (..),
   ContainerTestOptions (..),
-  OutputFormat (..),
+  TestOutputFormat (..),
   cliParser,
   mergeOpts,
   subcommand,
@@ -24,11 +24,14 @@ import App.Fossa.Config.Container.Common (
   imageTextArg,
  )
 import App.Fossa.Config.EnvironmentVars (EnvVars)
+import App.Fossa.Config.Test (TestOutputFormat (TestOutputJson, TestOutputPretty), defaultOutputFmt, testOutputFormatList, validateOutputFormat)
 import App.Types (OverrideProject (OverrideProject))
 import Control.Effect.Diagnostics (Diagnostics, Has)
+import Control.Monad (when)
 import Control.Timeout (Duration (Seconds))
 import Data.Aeson (ToJSON (toEncoding), defaultOptions, genericToEncoding)
 import Data.Text (Text)
+import Effect.Logger (Logger, logWarn, vsep)
 import Fossa.API.Types (ApiOpts)
 import GHC.Generics (Generic)
 import Options.Applicative (
@@ -40,10 +43,12 @@ import Options.Applicative (
   flag,
   help,
   info,
+  internal,
   long,
   option,
   optional,
   progDesc,
+  strOption,
  )
 
 subcommand :: (ContainerTestOptions -> a) -> Mod CommandFields a
@@ -54,18 +59,10 @@ subcommand f =
         progDesc "Check for issues from FOSSA and exit non-zero when issues are found"
     )
 
-data OutputFormat
-  = TestOutputPretty
-  | TestOutputJson
-  deriving (Eq, Ord, Show, Generic)
-
-instance ToJSON OutputFormat where
-  toEncoding = genericToEncoding defaultOptions
-
 data ContainerTestConfig = ContainerTestConfig
   { apiOpts :: ApiOpts
   , timeoutDuration :: Duration
-  , outputFormat :: OutputFormat
+  , outputFormat :: TestOutputFormat
   , testImageLocator :: ImageText
   , testRevisionOverride :: OverrideProject
   , testDockerHost :: Text
@@ -79,7 +76,8 @@ instance ToJSON ContainerTestConfig where
 data ContainerTestOptions = ContainerTestOptions
   { testCommons :: CommonOpts
   , containerTestTimeout :: Maybe Int
-  , containerTestOutputType :: OutputFormat
+  , deprecatedContainerTestOutputType :: TestOutputFormat
+  , containerTestOutputFmt :: Maybe String
   , containerTestImage :: ImageText
   }
 
@@ -88,11 +86,12 @@ cliParser =
   ContainerTestOptions
     <$> commonOpts
     <*> optional (option auto (long "timeout" <> help "Duration to wait for build completion (in seconds)"))
-    <*> flag TestOutputPretty TestOutputJson (long "json" <> help "Output issues as json")
+    <*> flag TestOutputPretty TestOutputJson (long "json" <> help "Output issues as json" <> internal)
+    <*> optional (strOption (long "format" <> help ("Output the report in the specified format. Currently available formats: (" <> testOutputFormatList <> ")")))
     <*> imageTextArg
 
 mergeOpts ::
-  Has Diagnostics sig m =>
+  (Has Diagnostics sig m, Has Logger sig m) =>
   Maybe ConfigFile ->
   EnvVars ->
   ContainerTestOptions ->
@@ -104,10 +103,30 @@ mergeOpts cfgfile envvars ContainerTestOptions{..} = do
       revOverride =
         collectRevisionOverride cfgfile $
           OverrideProject (optProjectName testCommons) (optProjectRevision testCommons) Nothing
+
+  -- if the non-default value is present for flag, user is using deprecatedJsonFlag
+  when (deprecatedContainerTestOutputType /= defaultOutputFmt) $ do
+    logWarn $
+      vsep
+        [ "DEPRECATION NOTICE"
+        , "=================="
+        , "--json flag is now deprecated for `fossa container test` command."
+        , ""
+        , "Please use: "
+        , "   `--format json` instead."
+        , ""
+        , "In future, usage of --json may result in fatal error."
+        ]
+
+  testOutputFormat <-
+    validateOutputFormat
+      (deprecatedContainerTestOutputType == TestOutputJson)
+      containerTestOutputFmt
+
   ContainerTestConfig
     <$> apiopts
     <*> pure timeout
-    <*> pure containerTestOutputType
+    <*> pure testOutputFormat
     <*> pure containerTestImage
     <*> pure revOverride
     <*> collectDockerHost envvars